feat(QRCodeCameraScanner.vue): add logic to determine if there are multiple cameras available and update internal state
The changes in this commit add logic to determine if there are multiple cameras available. This is done by checking the length of the `devices` array. If the length is greater than 1, the `hasMultipleCameras` value is set to `true`. This allows the component to know if there are multiple cameras available and update its internal state accordingly.